统一MVI架构
