+++ app/src/main/java/com/laomuji666/compose/navigation/NavigationHost.kt
@@ -6,30 +6,30 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.laomuji666.compose.core.ui.safePopBackStack
-import com.laomuji666.compose.feature.biometric.composeBiometricScreen
-import com.laomuji666.compose.feature.biometric.navigateToBiometricScreen
-import com.laomuji666.compose.feature.chat.chat.composeChatScreen
-import com.laomuji666.compose.feature.chat.chat.navigateToChatScreen
-import com.laomuji666.compose.feature.chat.composeAiChatScreen
-import com.laomuji666.compose.feature.chat.navigateToAiChatScreen
-import com.laomuji666.compose.feature.date.composeDateScreen
-import com.laomuji666.compose.feature.date.navigateToDateScreen
-import com.laomuji666.compose.feature.demo.RouteDemoScreen
-import com.laomuji666.compose.feature.demo.composeDemoScreen
-import com.laomuji666.compose.feature.firebase.composeFirebaseScreen
-import com.laomuji666.compose.feature.firebase.navigateToFirebaseScreen
-import com.laomuji666.compose.feature.http.composeHttpScreen
-import com.laomuji666.compose.feature.http.navigateToHttpScreen
-import com.laomuji666.compose.feature.painter.composePainterScreen
-import com.laomuji666.compose.feature.painter.navigateToPainterScreen
+import com.laomuji666.compose.feature.biometric.BiometricScreenRoute.composeBiometricScreen
+import com.laomuji666.compose.feature.biometric.BiometricScreenRoute.navigateToBiometricScreen
+import com.laomuji666.compose.feature.chat.AiChatScreenRoute.composeAiChatScreen
+import com.laomuji666.compose.feature.chat.AiChatScreenRoute.navigateToAiChatScreen
+import com.laomuji666.compose.feature.chat.chat.ChatScreenRoute.Companion.composeChatScreen
+import com.laomuji666.compose.feature.chat.chat.ChatScreenRoute.Companion.navigateToChatScreen
+import com.laomuji666.compose.feature.date.DateScreenRoute.composeDateScreen
+import com.laomuji666.compose.feature.date.DateScreenRoute.navigateToDateScreen
+import com.laomuji666.compose.feature.demo.DemoScreenRoute
+import com.laomuji666.compose.feature.demo.DemoScreenRoute.composeDemoScreen
+import com.laomuji666.compose.feature.firebase.FirebaseScreenRoute.composeFirebaseScreen
+import com.laomuji666.compose.feature.firebase.FirebaseScreenRoute.navigateToFirebaseScreen
+import com.laomuji666.compose.feature.http.HttpScreenRoute.composeHttpScreen
+import com.laomuji666.compose.feature.http.HttpScreenRoute.navigateToHttpScreen
+import com.laomuji666.compose.feature.painter.PainterScreenRoute.composePainterScreen
+import com.laomuji666.compose.feature.painter.PainterScreenRoute.navigateToPainterScreen
 import com.laomuji666.compose.feature.scroll.composeNestedScrollConnectionScreen
 import com.laomuji666.compose.feature.scroll.composeNestedScrollDispatcherScreen
 import com.laomuji666.compose.feature.scroll.navigateToNestedScrollConnectionScreen
 import com.laomuji666.compose.feature.scroll.navigateToNestedScrollDispatcherScreen
-import com.laomuji666.compose.feature.webview.composeWebViewScreen
-import com.laomuji666.compose.feature.webview.navigateToWebViewScreen
-import com.laomuji666.compose.feature.youtubedl.composeYoutubeDLScreen
-import com.laomuji666.compose.feature.youtubedl.navigateToYoutubeDLScreen
+import com.laomuji666.compose.feature.webview.WebViewScreenRoute.Companion.composeWebViewScreen
+import com.laomuji666.compose.feature.webview.WebViewScreenRoute.Companion.navigateToWebViewScreen
+import com.laomuji666.compose.feature.youtubedl.YoutubeDLScreenRoute.composeYoutubeDLScreen
+import com.laomuji666.compose.feature.youtubedl.YoutubeDLScreenRoute.navigateToYoutubeDLScreen
 
 /**
  * 导航控制器
@@ -38,7 +38,7 @@ import com.laomuji666.compose.feature.youtubedl.navigateToYoutubeDLScreen
 @Composable
 fun NavigationHost(
     navHostController: NavHostController,
-    startDestination:Any = RouteDemoScreen,
+    startDestination:Any = DemoScreenRoute,
     animTime:Int = 300
 ){
     NavHost(
