--- feature/http/src/main/kotlin/com/laomuji666/compose/feature/http/HttpNavigation.kt
@@ -1,25 +1,0 @@
-package com.laomuji666.compose.feature.http
-
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavHostController
-import androidx.navigation.NavOptions
-import androidx.navigation.compose.composable
-import com.laomuji666.compose.core.ui.navOptionsPushBack
-import kotlinx.serialization.Serializable
-
-@Serializable
-private data object RouteHttpScreen
-
-fun NavHostController.navigateToHttpScreen(navOptions: NavOptions = navOptionsPushBack()){
-    navigate(RouteHttpScreen, navOptions)
-}
-
-fun NavGraphBuilder.composeHttpScreen(
-    onBackClick:()->Unit
-){
-    composable<RouteHttpScreen>{
-        HttpScreen(
-            onBackClick = onBackClick
-        )
-    }
-}
