+++ feature/http/src/main/kotlin/com/laomuji666/compose/feature/http/HttpScreen.kt
@@ -103,7 +103,7 @@ private fun HttpScreenSlot(
 
 @Preview
 @Composable
-fun PreviewHttpScreen(){
+fun PreviewHttpScreenUi(){
     QuicklyTheme {
         HttpScreenUi(
             onBackClick = {},
