+++ feature/painter/src/main/kotlin/com/laomuji666/compose/feature/painter/PainterScreen.kt
@@ -236,7 +236,7 @@ private fun DrawScope.drawPathData(
 
 @Preview
 @Composable
-private fun PreviewPainterScreen() {
+private fun PreviewPainterScreenUi() {
     QuicklyTheme {
         PainterScreen(viewModel = PainterScreenViewModel())
     }
